+++ lib/crud/delete.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class DeleteData extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Colors.teal,
+        title: Text('Delete'),
+        elevation: 1.0,
+      ),
+      body: Center(
+        child: Text('DELETE DATA HERE'),
+      ),
+    );
+  }
+}
