+++ lib/crud/updata_data.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class Update extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Colors.teal,
+        title: Text('Update'),
+        elevation: 1.0,
+      ),
+      body: Center(
+        child: Text('UPDATE DATA HERE'),
+      ),
+    );
+  }
+}
